Update Configure-EnterpriseCA.ps1
There are wrong characters in the first five parameters of the Install-AdcsCertificationAuthority cmdlet. Instead of the dash character we have a diamond with a question mark in VS Code and Notepad++. Also when I hit edit in Github to propose the changes, the following message appeared:  We’ve detected the file encoding as windows-1252. When you commit changes we will transcode it to UTF-8. The character displayed in GitHub's edit more is '–' instead '-'. I fixed it and proposing the change to be incorporated in this branch.

With regards